+++ server/src/server.ts
@@ -9,8 +9,6 @@ import { TextDocument } from 'vscode-languageserver-textdocument';
 import { VizSymbol } from "./vizsymbol";
 import * as data from './viz_completions.json';
 import * as vizevent from './vizevent_completions.json';
-import * as net from 'net'
-import { ClientRequest } from 'http';
 
 
 
@@ -41,8 +39,8 @@ connection.onInitialize((params): ls.InitializeResult => {
 	//Initialize built in symbols
 	if (symbolCache["builtin"] == null) GetBuiltinSymbols();
 	if (symbolCache["builtin_events"] == null) GetVizEvents();
-	
-	
+
+
 
 	return {
 		capabilities: {
@@ -57,18 +55,10 @@ connection.onInitialize((params): ls.InitializeResult => {
 				resolveProvider: true,
 				triggerCharacters: ['.']
 			},
-			executeCommandProvider: {
-				commands: [
-					'vizscript.compile.setcurrentscene',
-					'vizscript.test'
-				]
-			},
-
 		}
 	
 	}
 });
-
 
 connection.onInitialized(() => {
 	if (hasConfigurationCapability) {
@@ -162,74 +152,6 @@ let sceneId = "";
 let containerId = "";
 let scriptId = "";
 
-connection.onExecuteCommand(async (params) => {
-
-	if(params.command == "vizscript.compile.setcurrentscene"){
-		let port = settings.compiler.hostPort;
-		let host = settings.compiler.hostName;
-
-		const socket = net.createConnection({ port: port, host: host}, () => {
-			// 'connect' listener.
-			//connection.console.log('Connected to Viz Engine!');
-			//connection.window.showInformationMessage("Connected to Viz Engine on " + host + ":" + port );
-			socket.write('1 MAIN IS_ON_AIR ' + String.fromCharCode(0));
-		});
-
-		let text = "";
-		text = documents.get(documentUri).getText();
-
-		socket.on('data', (data) => {
-			let message = data.toString().replace(String.fromCharCode(0), '')
-
-			let answer = GetRegexResult(message, /^([^\s]+)(\s?)(.*)/gi)
-			
-			if(answer[1] == "1"){
-				//connection.console.log("Viz Engine is OnAir");
-				socket.write('3 MAIN_SCENE*SCRIPT*PLUGIN*SOURCE_CODE SET ' + text + ' ' + String.fromCharCode(0));
-				
-			}else if(answer[1] == "3"){
-				//connection.console.log(answer[3]);
-				socket.write('4 MAIN_SCENE*SCRIPT*PLUGIN COMPILE ' + String.fromCharCode(0));
-			}else if(answer[1] == "4"){
-				//connection.console.log(answer[3]);
-				socket.write('5 MAIN_SCENE*SCRIPT*PLUGIN*COMPILE_STATUS GET ' + String.fromCharCode(0));
-			}else if(answer[1] == "5"){
-				//connection.console.log(answer[3]);
-
-				let error = GetRegexResult(answer[3], /\{(.*?)(\((.*)\))\}/gi)
-				if(error != undefined){
-					let rangesplit = error[3].split("/");
-					let line = parseInt(rangesplit[0]);
-					let char = parseInt(rangesplit[1]);
-					let range = ls.Range.create(line-1, char-1, line-1, char);
-
-					DisplayDiagnostics(documentUri,range,error[1]);
-					connection.window.showErrorMessage(scriptType + " script: Compilation failed: " + error[1])
-				}else{
-					connection.window.showInformationMessage(scriptType + " script: Compilation successful! No errors.")
-				}
-
-				socket.end();
-			}
-			
-		});
-
-		socket.on('error', () => {
-			connection.window.showErrorMessage("Not able to connect to Viz Engine " + host + ":" + port);
-			sceneId = "";
-			containerId = "";
-		});
-
-
-		socket.on('end', () => {
-			
-			//connection.console.log('Disconnected Viz Engine');
-			//connection.window.showInformationMessage("Disconnected from Viz Engine");
-		});
-	}
-	
-});
-
 
 function cleanPendingValidation(textDocument: ls.TextDocument): void {
 	const request = pendingValidationRequests[textDocument.uri];
@@ -245,8 +167,10 @@ function triggerValidation(textDocument: ls.TextDocument): void {
 		delete pendingValidationRequests[textDocument.uri];
 		RefreshDocumentsSymbols(textDocument.uri);
 		SetScriptType(textDocument.languageId);
-		if (settings.compiler.liveSyntaxChecking){
-			connection.sendNotification('requestCompile');
+		if(settings != null){
+			if (settings.compiler.liveSyntaxChecking){
+				connection.sendNotification('requestCompile');
+			}
 		}
 	}, validationDelayMs);
 }
@@ -697,6 +621,8 @@ connection.onCompletion((params: ls.CompletionParams, cancelToken: ls.Cancellati
 	if (GetRegexResult(line, /^[ \t]*dim[ \t]+([a-zA-Z0-9\-\_\,]+)[ \t]*([as]+)?$/gi) != null) return;// No sugggestions when declaring variables
 	if (GetRegexResult(line, /^[ \t]*function[ \t]+([a-zA-Z0-9\-\_\,]+)$/gi) != null) return;// No suggestions when declaring functions
 	if (GetRegexResult(line, /^[ \t]*end[ \t]+([a-zA-Z0-9\-\_\,]*)$/gi) != null) return;// No suggestions for ending sub or function
+	if (GetRegexResult(line, /^[if|elseif]*(.*)[ \t]+[then]+$/gi) != null) return;// No suggestions at end of if sentence
+	
 	if (GetRegexResult(line, /^[ \t]*sub[ \t]+On?$/gi) != null) return SelectBuiltinEventCompletionItems();// Only event suggestions when declaring submethod
 
 	if (GetRegexResult(line, /\=[\s]*([^\=\.\)]+)$/gi) != null) // Suggestions after "="
@@ -1858,6 +1784,7 @@ function DisplayDiagnostics(uri: string, range: ls.Range, message: string): void
 	diagnostics.push(diagnostic);
 
 	connection.sendDiagnostics({ uri, diagnostics });
+
 }
 
 function ClearDiagnostics(uri: string) {
@@ -1875,14 +1802,17 @@ connection.onRequest('getVizConnectionInfo', () => {
 });
 
 connection.onRequest('showDiagnostics', (vizReply: string) => {
-	let error = GetRegexResult(vizReply, /\{(.*?)(\((.*)\))\}/gi)
-	if(error != undefined){
+	let error = GetRegexResult(vizReply, /\{(.*?)(\((.*)\))?\}/gi)
+	if((error != undefined) && (error.length > 2)){
 		let rangesplit = error[3].split("/");
 		let line = parseInt(rangesplit[0]);
 		let char = parseInt(rangesplit[1]);
 		let range = ls.Range.create(line-1, char-1, line-1, char);
 		DisplayDiagnostics(documentUri,range,error[1])
-		return error[1] + " " + error[3]
+		connection.window.showErrorMessage(error[1])
+		return error[3] 
+	} else{
+		return error[1]
 	}
 })